// src/types/legacy-api-types.ts
export interface components {
    schemas: {
// ...
        OrderBooking: {
            /** Format: uuid */
            id: string;
            /** Format: uuid */
            tenant_id: string;
            /** Format: uuid */
            employee_id: string;
            /** Format: uuid */
            order_id: string;
            /** Format: uuid */
            activity_id?: string | null;
            /** Format: date */
            booking_date: string;
            /** @example 480 */
            time_minutes: number;
            description?: string | null;
            /**
             * @example manual
             * @enum {string}
             */
            source: "manual" | "auto" | "import";
            /** Format: date-time */
            created_at?: string;
            /** Format: date-time */
            updated_at?: string;
            /** Format: uuid */
            created_by?: string | null;
            /** Format: uuid */
            updated_by?: string | null;
        };
// ...
    };
// ...
}